data_pipeline/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pymupdf>=1.23.0
spacy>=3.7.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
